Refactor(DI): removed "manual" package, moved book to "model" package, added SimpleContainer class
