+++ src/main/java/org/example/di/container/SimpleContainer.java
@@ -1,0 +1,4 @@
+package org.example.di.container;
+
+public class SimpleContainer {
+}
